+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Thanks to @praksb, @ntotten and @vazexqi for their help on getting jorje fixed.
 - Workaround for certain cases of AST verification failing ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/38)).
 - Fix overlapping node with comment when comment contains special characters.
 - Fix awkward breaks for long method call chain ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/53)),
+- Remove breaks in Map types ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/57)).
 
 ## 1.0.0-alpha.9
 - Fix trailing comments after class names not being printed ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/20)).